Added the drawer
