+++ package.json
@@ -58,6 +58,7 @@
     "react-native-unordered-list": "^1.0.2",
     "react-native-web": "~0.11.7",
     "react-navigation": "~4.0.10",
+    "react-navigation-drawer": "^2.4.11",
     "react-navigation-stack": "~1.10.3",
     "react-navigation-tabs": "~2.6.2",
     "yarn": "^1.22.4"
